+++ src/test/java/io/dangernoodle/grt/utils/RepositoryDeserlizationTest.java
@@ -41,7 +41,6 @@ import static io.dangernoodle.RepositoryAsserts.verifyWiki;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.notNullValue;
-import static org.hamcrest.Matchers.nullValue;
 
 import java.io.IOException;
 import java.util.Arrays;
@@ -159,7 +158,7 @@ public class RepositoryDeserlizationTest
     private void thenPluginIsCorrect()
     {
         assertThat(repository.getPlugin("jenkins"), notNullValue());
-        assertThat(repository.getPlugin("doesnotexist"), nullValue());
+        assertThat(repository.getPlugin("doesnotexist"), equalTo(JsonObject.NULL));
     }
 
     private void thenPluginsIsCorrect()